fix: remove duplicated formatting, now have formatUtils
